remove config (debug) and rearrange files
