--- src/config.ts
@@ -1,5 +1,0 @@
-export let DEBUG = false;
-
-export function setDebug(value: boolean) {
-    DEBUG = value;
-}
