+++ src/parser/parse-pdfs.ts
@@ -1,0 +1,31 @@
+import {parsers, ParserType} from './all-parsers';
+import {ParsedOutput} from './parsed-output';
+
+export type StatementPdf<SelectedParser extends ParserType = ParserType> = {
+    parserInput: Parameters<typeof parsers[SelectedParser]['parser']>[0];
+    type: SelectedParser;
+};
+
+export type ParsedPdf<SelectedParser extends ParserType = ParserType> =
+    StatementPdf<SelectedParser> & {
+        data: ParsedOutput;
+    };
+
+export async function parsePdfs(
+    pdfs: Readonly<Readonly<StatementPdf>[]>,
+    debug = false,
+): Promise<Readonly<Readonly<ParsedPdf>[]>> {
+    const parsedPdfs: Readonly<Readonly<ParsedPdf>[]> = await Promise.all(
+        pdfs.map(async (pdf) => {
+            return {
+                ...pdf,
+                data: await parsers[pdf.type].parser({
+                    debug,
+                    ...pdf.parserInput,
+                }),
+            };
+        }),
+    );
+
+    return parsedPdfs;
+}
