+++ src/parser/parser-function.ts
@@ -9,6 +9,8 @@ export type ParseFunction<
 export type ParseFunctionInputs<ParserOptions extends object | undefined = undefined> = {
     /** FilePath Path of pdf to read */
     filePath: string;
+    /** Set to true to turn on additional logging */
+    debug?: boolean;
     /** Parser options that vary based on the parser being used */
-    parserOptions?: Partial<CombineWithBaseParserOptions<ParserOptions>>;
+    inputParserOptions?: Partial<CombineWithBaseParserOptions<ParserOptions>>;
 };
